Coin Activity Bottom Shee 11.25.2021

+++ app/src/main/java/com/digitalartsplayground/fantasycrypto/models/CandleStickData.java
@@ -6,9 +6,11 @@ import com.google.gson.annotations.SerializedName;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CandleStickData extends ArrayList<List<Double>> {
+public class CandleStickData extends ArrayList<List<Float>> {
 
-    private ArrayList<List<Double>> getCandleStickData(){
+    private ArrayList<List<Float>> getCandleStickData(){
         return this;
     }
+
+
 }
